% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfmat.r
\name{cpumat-class}
\alias{cpumat-class}
\alias{hdfmatR6}
\title{cpumat class}
\description{
Storage and methods for CPU matrix data.
}
\details{
Data is held in an external pointer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{hdfmatR6$new()}}
\item \href{#method-fill}{\code{hdfmatR6$fill()}}
\item \href{#method-read}{\code{hdfmatR6$read()}}
\item \href{#method-scale}{\code{hdfmatR6$scale()}}
\item \href{#method-set_diag}{\code{hdfmatR6$set_diag()}}
\item \href{#method-crossprod}{\code{hdfmatR6$crossprod()}}
\item \href{#method-clone}{\code{hdfmatR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$new(file, name, nrows, ncols, type = "double")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{File to store data in.}

\item{\code{name}}{Dataset name on disk.}

\item{\code{nrows, ncols}}{The dimension of the matrix.}

\item{\code{type}}{Storage type for the matrix. Should be one of 'int', 'float', or 'double'.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Class initializer.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill"></a>}}
\if{latex}{\out{\hypertarget{method-fill}{}}}
\subsection{Method \code{fill()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$fill(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale"></a>}}
\if{latex}{\out{\hypertarget{method-scale}{}}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$scale(v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_diag"></a>}}
\if{latex}{\out{\hypertarget{method-set_diag}{}}}
\subsection{Method \code{set_diag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$set_diag(v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crossprod"></a>}}
\if{latex}{\out{\hypertarget{method-crossprod}{}}}
\subsection{Method \code{crossprod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$crossprod(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
