% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprod_ooc.r
\name{crossprod_ooc}
\alias{crossprod_ooc}
\title{crossprod_ooc}
\usage{
crossprod_ooc(x, file, name = "crossprod")
}
\arguments{
\item{x}{The input matrix. Should be in double precision.}

\item{file}{Name of the file to use for the out-of-core storage.}

\item{name}{The dataset name within the HDF5 file.}
}
\value{
Returns an hdfmat object.
}
\description{
Out-of-core crossproduct. Useful when the number of columns is very large.
}
