% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfmat.r
\name{hdfmat-class}
\alias{hdfmat-class}
\alias{hdfmatR6}
\title{hdfmat class}
\description{
Storage and methods for matrix data in an HDF5 file.
}
\details{
Data is held in an external pointer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{hdfmatR6$new()}}
\item \href{#method-close}{\code{hdfmatR6$close()}}
\item \href{#method-print}{\code{hdfmatR6$print()}}
\item \href{#method-fill}{\code{hdfmatR6$fill()}}
\item \href{#method-read}{\code{hdfmatR6$read()}}
\item \href{#method-scale}{\code{hdfmatR6$scale()}}
\item \href{#method-set_diag}{\code{hdfmatR6$set_diag()}}
\item \href{#method-crossprod}{\code{hdfmatR6$crossprod()}}
\item \href{#method-eigen}{\code{hdfmatR6$eigen()}}
\item \href{#method-svd}{\code{hdfmatR6$svd()}}
\item \href{#method-clone}{\code{hdfmatR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$new(open, file, name, nrows, ncols, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{open}}{Are you working with an existing matrix? If so, \code{nrows},
\code{ncols}, and \code{type} are ignored (and intuited).}

\item{\code{file}}{File to store data in.}

\item{\code{name}}{Dataset name on disk.}

\item{\code{nrows, ncols}}{The dimension of the matrix.}

\item{\code{type}}{Storage type for the matrix. Should be one of 'int', 'float', or 'double'.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Class initializer.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$close()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Closes the file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print some basic info about an hdfmat object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill"></a>}}
\if{latex}{\out{\hypertarget{method-fill}{}}}
\subsection{Method \code{fill()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$fill(x, row_offset = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The input matrix. Must be double or float.}

\item{\code{row_offset}}{The number of rows to skip. Useful when building the
hdfmat in chunks.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fill an hdfmat matrix (or part of it) with the input matrix. The number
of rows of the input must be less than or equal to the number of rows
allocated for the hdfmat, and the number of columns must be equal.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$read()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Read an hdfmat-stored matrix into memory. Really only meant for testing.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale"></a>}}
\if{latex}{\out{\hypertarget{method-scale}{}}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$scale(v)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v}}{Scalar. Fundamental type can be double, float, or int.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Scale (multiply) all values of an hdfmat-stored matrix by the input
scalar.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_diag"></a>}}
\if{latex}{\out{\hypertarget{method-set_diag}{}}}
\subsection{Method \code{set_diag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$set_diag(v)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v}}{Scalar. Fundamental type can be double, float, or int.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the diagonal of an hdfmat-stored matrix to the input scalar.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crossprod"></a>}}
\if{latex}{\out{\hypertarget{method-crossprod}{}}}
\subsection{Method \code{crossprod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$crossprod(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Input matrix. Fundamental type can be double, float, or int.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Calculate the crossproduct of an input matrix with result stored in an
hdfmat. Useful when the number of columns of the input is very large.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eigen"></a>}}
\if{latex}{\out{\hypertarget{method-eigen}{}}}
\subsection{Method \code{eigen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$eigen(k = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{The number of Lanczos iterations.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Compute approximations to the eigenvalues of a square symmetric
hdfmat-stored matrix using the Lanczos method. The matrix is not checked
for symmetry.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-svd"></a>}}
\if{latex}{\out{\hypertarget{method-svd}{}}}
\subsection{Method \code{svd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$svd(k = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{The number of Lanczos iterations.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Compute approximations to the singular values of a rectangular
hdfmat-stored matrix using the Lanczos method.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hdfmatR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
