% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfmat.r
\name{hdfmat}
\alias{hdfmat}
\title{hdfmat}
\usage{
hdfmat(file, name, nrows, ncols, type = "double", compression = 0L)
}
\arguments{
\item{file}{File to store data in.}

\item{name}{Dataset name on disk.}

\item{nrows, ncols}{The dimension of the matrix.}

\item{type}{Storage type for the matrix. Should be one of 'int', 'float', or
'double'.}

\item{compression}{The compression level, an integer from 0 (no compression)
to 9 (highest compression). Run-time performance degrades with increased
compression levels.}
}
\value{
An hdfmat class object.
}
\description{
Constructor for hdfmat objects.
}
