% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfmat.r
\name{hdfmat_open}
\alias{hdfmat_open}
\title{hdfmat_open}
\usage{
hdfmat_open(file, name)
}
\arguments{
\item{file}{File to store data in.}

\item{name}{Dataset name on disk.}
}
\value{
An hdfmat class object.
}
\description{
Constructor for hdfmat objects.
}
