% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfmat.r
\name{hdfmat}
\alias{hdfmat}
\title{hdfmat}
\usage{
hdfmat(file, name, nrows, ncols, type = "double")
}
\arguments{
\item{nrows, ncols}{The dimensions of the matrix.}

\item{type}{Storage type for the matrix. Should be one of 'int', 'float', or 'double'.}
}
\value{
A cpumat class object.
}
\description{
Constructor for cpumat objects.
}
\details{
Data is held in an external pointer.
}
